requirements.txt
python-dotenv==0.15.0
opencv-python==4.5.1.48
requests==2.25.1
Pillow==8.1.0
imutils==0.5.4
moviepy==1.0.3
google-api-python-client==1.9.3
google-auth-oauthlib==0.4.1
google-auth-httplib2==0.0.3
python-dateutil==2.8.1